move Future to util package
